Remove custom date argument.
Arguments and Operators shouldn't need to validate/transform their inputs to work.  Instead, an Input should only provide valid values, and when creating Switch conditions that use Operators which only operate on certain types (i.e. Dates), then that is the responsibility of the creator of the Operator to pass in the correct value when used in a condition where the Input's argument returns a date.