Change Boolean argument to hash a specific value, rather than its own.
This makes it work with the percentage operator, which hashes the argument.  For boolean arguments, its value is always "True" so it will always hash to the exact same thing; "1".  That won't work if you want the operator to apply to a percentage of those arguments, so the hash_value is used instead.